# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Realize your dreams by becoming a Fastpay Modern Shop Partner.",
    "version": "1.0",
    "title": "FASTPAY API,s",
    "contact": {
      "email": "link2mirfan@gmail.com"
    }
  },
  "host": "128.199.170.4:3008",
  "basePath": "/api",
  "tags": [
    {
      "name": "fastpay",
      "description": "Access to fastpay apis"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "paths": {
    "/mobile-credit": {
      "post": {
        "tags": [
          "fastpay"
        ],
        "summary": "Pay your mobile credit",
        "description": "",
        "operationId": "mobileCredit",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Body object that needs to be create transection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mobileCredit"
            }
          }
        ],
        "responses": {
          "200": {
            "description":"successful operation"
          },
          "400": {
            "description": "Reqired data is missing"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/pay-phone-bill": {
      "post": {
        "tags": [
          "fastpay"
        ],
        "summary": "Pay your phone bill",
        "description": "",
        "operationId": "payPhoneBill",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Body object that needs to be create transection",
            "required": true,
            "schema": {
              "$ref": "#/definitions/payPhoneBill"
            }
          }
        ],
        "responses": {
          "200": {
            "description":"successful operation"
          },
          "400": {
            "description": "Reqired data is missing"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    }
    
  },
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "mobileCredit": {
      "type": "object",
      "required": [
        "kode_produk",
        "no_hp"
      ],
      "properties": {
        "kode_produk": {
          "type": "string",
          "example": "I10H"
        },
        "no_hp": {
          "type": "string",
          "example": "085648889293"
        },
        "ref1": {
          "type": "string",
          "example": "ref1 value"
        }
      },
      "xml": {
        "name": "mobileCredit"
      }
    },
    "payPhoneBill": {
      "type": "object",
      "required": [
        "area_code",
        "phone_number",
        "ref2",
        "nominal"
      ],
      "properties": {
        "area_code": {
          "type": "string"
        },
        "phone_number": {
          "type": "string"
        },
        "nominal": {
          "type": "string"
        },
        "ref1": {
          "type": "string",
          "example": "ref1 value"
        },
        "ref2": {
          "type": "string",
          "example": "ref2 value"
        }
      },
      "xml": {
        "name": "mobileCredit"
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}